+++ src/main/java/raisetech/Student/management/Application.java
@@ -1,30 +1,13 @@
 package raisetech.Student.management;
 
-import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 @SpringBootApplication
-@RestController
-public class Application {
 
-	@Autowired
-	private StudentRepository repository;
+public class Application {
 
 	public static void main(String[] args) {
 		SpringApplication.run(Application.class, args);
-	}
-
-	@GetMapping("/studentList")
-	public List<Student> getStudentList() {
-		return repository.search();
-	}
-
-	@GetMapping("/studentCourseList")
-	public List<StudentCourse> getStudentCourseList() {
-		return repository.searchCourse();
 	}
 }
