+++ src/main/java/raisetech/Student/management/repository/StudentRepository.java
@@ -1,8 +1,10 @@
-package raisetech.Student.management;
+package raisetech.Student.management.repository;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
+import raisetech.Student.management.data.Student;
+import raisetech.Student.management.data.StudentCourse;
 
 @Mapper
 public interface StudentRepository {
@@ -11,5 +13,5 @@ public interface StudentRepository {
   List<Student> search();
 
   @Select("SELECT * FROM students_courses")
-  List<StudentCourse> searchCourse();
+  List<StudentCourse> searchStudentCourse();
 }
